Put together the first stable version of the app
Once again, the ES6 Modules powered by Webpack/NPM were rather
challenging to work with, given that the file structure in Vanilla JS
is completely up to the developer and there are only a few standards to
follow.

The moment modular JS was off and running, everything came down to
address each feature one by one, such as making the data persistent
and populating modals apart.